+++ src/fonts.ts
@@ -1,2 +1,6 @@
-export { GoogleFont, CustomFont } from "@cf-wasm/og/workerd";
-export { loadGoogleFont, createFontConfig } from "./fonts.shared";
+export {
+  GoogleFont,
+  CustomFont,
+  loadGoogleFont,
+  createFontConfig,
+} from "./fonts.shared";