+++ src/image-response.node.ts
@@ -1,9 +1,11 @@
-import { ImageResponse as CfImageResponse } from "@cf-wasm/og/node";
 import type { ReactNode } from "react";
 import { createImageResponseClass } from "./core/image-response";
+import { renderPng, renderSvg } from "./runtime/satori.node";
+import { cache } from "./cache";
 
-export { cache } from "@cf-wasm/og/node";
+export { cache };
 
 export const ImageResponse = createImageResponseClass<ReactNode>({
-  cfImageResponse: CfImageResponse,
+  renderSvg,
+  renderPng,
 });